pyproject.toml
[project]
name = "sse-delta-parser"
version = "0.2.0"
description = "SSE Delta Log Aggregator - Parse and combine fragmented SSE delta logs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "jinja2>=3.1.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    # Authentication
    "authlib>=1.3.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.27.0",
    # Settings
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    # Session middleware
    "itsdangerous>=2.1.0",
]

[tool.uv]
dev-dependencies = []
